fixing types & moving hard-coded values
