--- components/Case/types.d.ts
@@ -1,12 +1,0 @@
-export type Tech =
-  | 'next'
-  | 'typescript'
-  | 'css'
-  | 'react'
-  | 'html'
-  | 'motion'
-  | 'tailwind'
-  | 'github'
-  | 'git'
-  | 'node'
-  | 'supabase';
