+++ utils/icons.ts
@@ -1,0 +1,33 @@
+import GlobeIcon from '@/public/cases/globe.svg';
+import GithubIcon from '@/public/cases/github.svg';
+import RightArrow from '@/public/cases/arrowright.svg';
+import DefIcon from '@/public/cases/def.png';
+import DesignIcon from '@/public/cases/design.png';
+import PlanIcon from '@/public/cases/plan.png';
+import DevIcon from '@/public/cases/dev.png';
+import DeployIcon from '@/public/cases/deploy.png';
+import DefLightIcon from '@/public/cases/defLight.png';
+import DesignLightIcon from '@/public/cases/designLight.png';
+import PlanLightIcon from '@/public/cases/planLight.png';
+import DevLightIcon from '@/public/cases/devLight.png';
+import DeployLightIcon from '@/public/cases/deployLight.png';
+import ChallengeIcon from '@/public/cases/challenge.png';
+import SuccessIcon from '@/public/cases/success.png';
+
+export {
+  GlobeIcon,
+  GithubIcon,
+  RightArrow,
+  DefIcon,
+  DesignIcon,
+  PlanIcon,
+  DevIcon,
+  DeployIcon,
+  DefLightIcon,
+  DesignLightIcon,
+  PlanLightIcon,
+  DevLightIcon,
+  DeployLightIcon,
+  ChallengeIcon,
+  SuccessIcon,
+};
